+++ Health/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
             android:label="@string/app_name"
             android:supportsRtl="true"
             android:theme="@style/AppTheme"
-            android:backupAgent=".data.RealmBackupAgent"
+            android:backupAgent=".core.RealmContext$RealmBackupAgent"
             android:fullBackupOnly="true"
             android:fullBackupContent="true">
 
@@ -33,11 +33,11 @@
             </intent-filter>
         </activity>
 
-	<!-- For Clock testing. -->
+        <!-- For Clock testing. -->
         <activity
-            android:name=".ClockActivity"
-            android:label="@string/title_activity_clock"
-            android:theme="@style/AppTheme.NoActionBar"></activity>
+                android:name=".ClockActivity"
+                android:label="@string/title_activity_clock"
+                android:theme="@style/AppTheme.NoActionBar"/>
 
         <!-- For SleepModule to play audio in the background. -->
         <service
