+++ Health/src/main/java/com/sciencesquad/health/sleep/SleepModule.java
@@ -3,9 +3,9 @@ package com.sciencesquad.health.sleep;
 import android.util.Pair;
 import com.sciencesquad.health.R;
 import com.sciencesquad.health.core.BaseApp;
+import com.sciencesquad.health.core.DataContext;
 import com.sciencesquad.health.core.Module;
-import com.sciencesquad.health.data.DataContext;
-import com.sciencesquad.health.data.RealmContext;
+import com.sciencesquad.health.core.RealmContext;
 import com.sciencesquad.health.nutrition.NutritionModel;
 
 /**
